Add Projects class outside of actions
